Move Sections folder to inside Application component folders to keep it local
