+++ src/components/Application/Sections/SectionsList.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { List } from 'semantic-ui-react'
-import { SectionDetails } from '../../utils/types'
+import { SectionDetails } from '../../../utils/types'
 
 interface SectionsListProps {
   sections: SectionDetails[]
