+++ src/components/Application/Sections/index.ts
@@ -1,0 +1,4 @@
+import SectionsList from './SectionsList'
+import SectionsProgress from './SectionsProgress'
+
+export { SectionsList, SectionsProgress }
