--- src/components/Sections/index.ts
@@ -1,3 +1,0 @@
-import SectionsProgress from './SectionsProgress'
-
-export { SectionsProgress }
